pyproject.toml
[project]
name = "p-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiofile==3.8.8",
    "aiohttp==3.8.4",
    "aiomqtt==2.2.0",
    "asyncify==0.9.2",
    "bilibili-api-python==15.5.1",
    "borax==4.1.1",
    "cn2an==0.5.22",
    "dacite==1.8.1",
    "google-generativeai==0.8.1",
    "graiax-silkcoder==0.3.6",
    "huaweicloudsdkcore==3.1.56",
    "huaweicloudsdkimage==3.1.56",
    "huaweicloudsdksis==3.1.56",
    "humanize==4.12.3",
    "imagehash==4.3.1",
    "imageio==2.12.0",
    "inflection==0.5.1",
    "mako==1.1.5",
    "nest-asyncio==1.5.7",
    "py-mini-racer==0.6.0",
    "pychorus==0.1",
    "pydub==0.25.1",
    "pyncm-async==0.1.2",
    "pyzbar==0.1.8",
    "uvicorn>=0.35.0",
    "yiri-mirai==0.2.6.2",
    "zhconv==1.4.3",
]
